agent_backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
faiss-cpu>=1.9.0
pymongo>=4.10.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.28.0
numpy>=1.26.0
